\name{multivariate gamma functions}

\alias{mgamma}
\alias{lmgamma}
\alias{mdigamma}

\title{
Multivariate Gamma Functions
}

\description{
See Value.
}

\usage{
mgamma(x, D)
lmgamma(x, D)
mdigamma(x, D)
}

\arguments{
\item{x}{numeric.}
\item{D}{positive integer; the dimension of the function.}
}

\value{the multivariate gamma function, the logarithm of the multivariate gamma function, or the multivariate digamma function (the derivative of the logarithm of the multivariate gamma function) evaluated at \code{x} for dimension \code{D}.}

\author{
Daniel Dvorkin
}
