\name{constants}

\alias{constants}
\alias{LC_LOG2PI}
\alias{LC_EPS}
\alias{LC_LEPS}
\alias{LC_REPS}
\alias{LC_LREPS}
\alias{LC_EPSEPS}
\alias{LC_MAXSTDNORM}
\alias{LC_MINSTDNORM}
\alias{LC_ITER_MAX}
\alias{LC_TOPOLOGY}
\alias{LC_FAMILY}
\alias{LC_NONNEGFAM}

\docType{data}

\title{
``Constants'' Used in Program Control
}

\description{
Values used in controlling the executions of various functions in the \code{lcmix} package.
}

\details{
All ``constants'' should be treated as though they were actual constants, and \emph{never} reassigned.  For example, strenuously avoid statements such as \code{LC_EPS <- 1e-20}.
}

\value{
\item{LC_LOG2PI}{\eqn{\log(2\pi)}{log(2*pi)}, approximately 1.837877.}
\item{LC_EPS}{``epsilon,'' a number very near 0.  Defined here as \code{sqrt(.Machine$double.eps)}, or approximately 1.490116e-08 on most systems.}
\item{LC_LEPS}{the logarithm of \code{LC_EPS}.}
\item{LC_REPS}{``reverse epsilon,'' a number very near 1.  Defined here as \code{1-LC_EPS}.}
\item{LC_LREPS}{the logarithm of \code{LC_REPS}.}
\item{LC_EPSEPS}{a number even closer to 0 than \code{LC_EPS}, defined here as \eqn{10^{-300}}{1e-300}.  Use with caution as \code{LC_EPSEPS} will be indistinguishable from 0 under many circumstances.}
\item{LC_MAXSTDNORM (LC_MINSTDNORM)}{maximum (minimum) values for numbers supposed to have been generated by a standard normal distribution, defined here as 8 (-8).  Values greater (less) than this will have a standard normal density very near 0.}
\item{LC_ITER_MAX}{default maximum iterations for the EM algorithm as used in \code{mixmod} and \code{mdmixmod}, defined here as 10000.}
\item{LC_TOPOLOGY}{vector of allowable topology names for multiple-data mixture models fitted using \code{mdmixmod}, currently \code{"layered"} and \code{"chained".}}
\item{LC_FAMILY}{a list of which each element bears the name of an allowable distribution family for mixture models.  Each list element is itself a two-element list with elements \code{uni} and \code{multi}, corresponding to the names of the univariate and multivariate versions of the distribution family.  The current families, and univariate and multivariate names, are given in the following table.
	\tabular{lll}{
		\emph{family} \tab \emph{univariate} \tab \emph{multivariate} \cr
		normal        \tab norm              \tab mvnorm              \cr
		weibull       \tab weisd             \tab mvweisd             \cr
		gamma         \tab gamma             \tab mvgamma       
	}
}
\item{LC_NONNEGFAM}{a vector of names of distribution families which require data to have non-negative values, currently \code{"weibull"} and \code{"gamma"}.}
}

\keyword{datasets}
